+++ torch_timeseries/model/__init__.py
@@ -18,6 +18,7 @@ from .DiffusionTS import DiffusionTS
 from .TimeDiff import TimeDiff
 from .NsDiff import NsDiff
 from .TMDM import TMDM
+from .SegRNN import SegRNN
 from .diffusion_utils import GaussianDiffusion, make_beta_schedule, sinusoidal_embedding
 from .irregular import GRUD, mTAN, LatentODE, NeuralCDE, Raindrop
 
@@ -27,6 +28,7 @@ forecasting_models = [
     "PatchTST", "iTransformer",
     "TSMixer", "Crossformer", "SCINet", "TimesNet",
     "CATS", "FITS", "FreTS",
+    "SegRNN",
 ]
 
 generation_models = [